add same delete model to faculty, department, subject, direction, employee and group

+++ AVN/Controllers/FacultyController.cs
@@ -1,9 +1,11 @@
 ﻿using AVN.Automapper;
+using AVN.Data;
 using AVN.Data.UnitOfWorks;
 using AVN.Model.Entities;
 using AVN.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore;
 
 namespace AVN.Web.Controllers
 {
@@ -11,11 +13,13 @@ namespace AVN.Web.Controllers
     {
         private readonly IUnitOfWork unitOfWork;
         private readonly IMapper mapper;
+        private readonly AppDbContext context;
 
-        public FacultyController(IUnitOfWork unitOfWork, IMapper mapper)
+        public FacultyController(IUnitOfWork unitOfWork, IMapper mapper, AppDbContext context)
         {
             this.unitOfWork = unitOfWork;
             this.mapper = mapper;
+            this.context = context;
         }
 
         // GET: Faculty
@@ -107,10 +111,29 @@ namespace AVN.Web.Controllers
         // POST: Faculty/Delete/5
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var faculty = await unitOfWork.FacultyRepository.GetByIdAsync(id);
-            await unitOfWork.FacultyRepository.DeleteAsync(faculty);
-            await unitOfWork.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            try
+            {
+                var faculty = await context.Faculties.Include(d => d.Departments).FirstOrDefaultAsync(i => i.Id == id);
+                if (faculty == null)
+                {
+                    TempData["error"] =
+                        "Не удалось найти факультеты. Пожалуйста попробуйте позже, либо обратитесь к администратору.";
+                    return RedirectToAction("Index", "Faculty");
+                }
+
+                TempData["success"] = "Запись успешно удалена";
+                await unitOfWork.DepartmentRepository.DeleteRangeAsync(faculty.Departments);
+
+                await unitOfWork.FacultyRepository.DeleteAsync(faculty);
+                await unitOfWork.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (Exception ex)
+            {
+                TempData["error"] = $"Произошла внутренняя ошибка: {ex.Message}.  Пожалуйста попробуйте позже, либо обратитесь к администратору.";
+                return RedirectToAction("Index", "Faculty");
+            }
+            
         }
 
         public async Task<List<SelectListItem>> GetFaculties()
